# fix: improve startup performance of LeanClient