--- a/vscode-lean4/package.json
+++ b/vscode-lean4/package.json
@@ -2,7 +2,7 @@
 	"name": "lean4",
 	"displayName": "lean4",
 	"description": "Lean 4 language support for VS Code",
-	"version": "0.0.67",
+	"version": "0.0.68",
 	"publisher": "leanprover",
 	"engines": {
 		"vscode": "^1.57.0"