1. add optimization that skips the lake test on well known versions because it is too slow where the optimization is based on well known lean-toolchain version strings.  
1. Cache `lean --version` check results because we were doing 2 checks on each reload.

### Results

I put a console output that measures full lean server "restart" time following a `Select Toolchain` choice, so I can run release bits to get good measurements, and with this PR it is:

| Version      | Full restart time| Lake? |
| ----------- | ----------- | -------------|
| leanprover/lean4:nightly-2022-02-01| 658 ms | no |
| leanprover/lean4:nightly-2022-02-16 | 823 ms | yes |
| leanprover/lean4:4.0.0-m3 | 656 ms | no |

Previously without this PR the times are:

| Version      | Full restart time | Lake? |
| ----------- | ----------- |-------------|
| leanprover/lean4:nightly-2022-02-01| 1594 ms | no |
| leanprover/lean4:nightly-2022-02-16 | 1526 ms | yes |
| leanprover/lean4:4.0.0-m3 | 1637 ms | no |

So I think this shows a pretty healthy speed up.