@gebner 2022-02-16
> remove the lake test because it is too slow and instead switch to a test based on the lean-toolchain version string instead which is a lot faster.
> The Lean 4: Restart Server command goes from 2.7 seconds to about 1.6 seconds.

I agree that we should improve the startup performance, I've also noticed that there's a noticeable delay.  But I don't think it's *that* bad for me.

Can you post the runtime of the following commands on your machine?
```bash
lean --version             #  31ms
lake --version             #  58ms
lake serve </dev/null      # 645ms
lean --server </dev/null   # 378ms
```

If the numbers are similar on your machine, then we should just get rid of the `lake serve` check.  At the moment, this would cause a small regression where the server fails to start if the `lakefile.lean` contains syntax errors.  But we can just bug @tydeu to release a new lake version with https://github.com/leanprover/lake/pull/52 to fix this regression.