@tydeu 2022-02-16
> But we can just bug @tydeu to release a new lake version with [leanprover/lake#52](https://github.com/leanprover/lake/pull/52) to fix this regression.

Note that one can just check for a version greater than `3.0.0`, since any future version will include that fix.