@gebner 2022-02-16
> Note that one can just check for a version greater than 3.0.0, since any future version will include that fix.

But then find-references won't work until you release 3.0.1. :smile: