@tydeu 2022-02-16
@gebner what do you mean? The current version in the Lean repo is `3.1.0-pre`, so if the extension checks for versions greater than `3.0.0` it should activate with the current master but not for older nightlies using `3.0.0`. Right?