@lovettchris 2022-02-16
> Can you post the runtime of the following commands on your machine?
```
lean --version             # 78ms
lake --version             #  119ms
lake serve </dev/null      # 500ms
lean --server </dev/null   # 431ms
```