@lovettchris 2022-02-16
> Thanks, I missed the lake update yesterday. In that case everything is fine on the lake side and we just need to bump the version check here.

If I understand correctly you guys want me to change it from:
```typescript
 if (actual.compare('3.0.0') >= 0) {
```
to:
```typescript
 if (actual.compare('3.0.0') > 0) {
```
if so I optimistically pushed this change.